+++ src/app/classes/message.ts
