+++ src/app/components/cita/cancel-cita-popup.component.css
