+++ src/app/guards/logged.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
-import { UsersService } from './users.service';
+import { UsersService } from 'src/app/services/users.service';
 
 @Injectable({
   providedIn: 'root'
